+++ golem-component-service/src/lib.rs
@@ -14,7 +14,6 @@
 
 pub mod api;
 pub mod config;
-pub mod db;
 pub mod grpcapi;
 pub mod metrics;
 pub mod repo;
