+++ golem-component-service/src/service.rs
@@ -22,9 +22,9 @@ use golem_component_service_base::service::component_compilation::{
 use golem_service_base::config::ComponentStoreConfig;
 use golem_service_base::service::component_object_store;
 use std::sync::Arc;
+use golem_service_base::db;
 
 use crate::config::{ComponentServiceConfig, DbConfig};
-use crate::db;
 use crate::repo::component::{ComponentRepo, DbComponentRepo};
 
 #[derive(Clone)]
