+++ golem-service-base/src/config.rs
@@ -41,3 +41,38 @@ pub struct ComponentStoreLocalConfig {
     pub root_path: String,
     pub object_prefix: String,
 }
+
+
+
+#[derive(Clone, Debug, Deserialize)]
+#[serde(tag = "type", content = "config")]
+pub enum DbConfig {
+    Postgres(DbPostgresConfig),
+    Sqlite(DbSqliteConfig),
+}
+
+impl Default for DbConfig {
+    fn default() -> Self {
+        DbConfig::Sqlite(DbSqliteConfig {
+            database: "golem_component_service.db".to_string(),
+            max_connections: 10,
+        })
+    }
+}
+
+#[derive(Clone, Debug, Deserialize)]
+pub struct DbSqliteConfig {
+    pub database: String,
+    pub max_connections: u32,
+}
+
+#[derive(Clone, Debug, Deserialize)]
+pub struct DbPostgresConfig {
+    pub host: String,
+    pub database: String,
+    pub username: String,
+    pub password: String,
+    pub port: u16,
+    pub max_connections: u32,
+    pub schema: Option<String>,
+}
