+++ golem-service-base/src/db/mod.rs
@@ -88,7 +88,7 @@ pub async fn postgres_migrate(config: &DbPostgresConfig) -> Result<(), Box<dyn E
         return Err(format!("DB schema {} do not exists/was not created", schema).into());
     }
 
-    sqlx::migrate!("./db/migration/postgres")
+    sqlx::migrate!("db/migration/postgres")
         .run(&mut conn)
         .await?;
 
@@ -111,7 +111,7 @@ pub async fn sqlite_migrate(config: &DbSqliteConfig) -> Result<(), Box<dyn Error
     info!("DB migration: sqlite://{}", config.database);
     let conn_options = SqliteConnectOptions::from(config);
     let mut conn = SqliteConnection::connect_with(&conn_options).await?;
-    sqlx::migrate!("./db/migration/sqlite")
+    sqlx::migrate!("db/migration/sqlite")
         .run(&mut conn)
         .await?;
     let _ = conn.close().await;
