+++ golem-worker-service-base/Cargo.toml
@@ -44,6 +44,14 @@ serde_json = { workspace = true }
 serde_yaml = { workspace = true }
 strum = { workspace = true }
 strum_macros = { workspace = true }
+sqlx = { version = "0.7", features = [
+    "runtime-tokio",
+    "sqlite",
+    "postgres",
+    "uuid",
+    "migrate",
+    "chrono",
+] }
 tap = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
